Add missing --project flag
One of the gcloud commands in the service key-creation script was missing
the `--project` flag, meaning it could fail (when no gcloud default project is set)
or try to perform an operation on the wrong project (when the default is different
then the user-provided arg). Fixed by adding this flag. Tested on a machine with
no default project set, new behavior is correct.