Rewrite generate.sh for current lib versions
Without this rewrite, the old script failed with:

Generating deepcopy funcs
F0211 10:52:30.681397   71752 deepcopy.go:885] Hit an unsupported type invalid type for invalid type, from github.com/kubecost/cluster-turndown/pkg/apis/turndownschedule/v1alpha1.TurndownSchedule

Initially I found this issue:
https://github.com/operator-framework/operator-sdk/issues/1854 and tried
to use the GOROOT fix, but it didn't work. I ended up getting (with
set -x):

$ GOROOT=/usr/bin bash generate.sh
...
+ bash ./vendor/k8s.io/code-generator/generate-groups.sh all github.com/kubecost/cluster-turndown/pkg/generated github.com/kubecost/cluster-turndown/pkg/apis turndownschedule:v1alpha1 --output-base . --go-header-file hack/custom-boilerplate.go.txt
/home/delta/go/pkg/mod/golang.org/x/tools@v0.0.0-20210106214847-113979e3529a/internal/imports/imports.go:12:2: package bufio is not in GOROOT (/usr/bin/src/bufio)
/home/delta/go/pkg/mod/github.com/spf13/pflag@v1.0.5/flag.go:102:2: package bytes is not in GOROOT (/usr/bin/src/bytes)
... etc.

I opted to rewrite the generate.sh script in the image of the official
K8s sample-controller, with some additions based on the previous
generate.sh described in comments. I'm not sure if the go mod vendor
trick in the old generate.sh, which purportedly got around some trouble
caused when the code wasn't yet generated, will be necessary to
resurrect in the future if we update the API version. I'll document this
in the PR as well, so hopefully someone can find this with some
spelunking if the problem arises again.

After the rewrite, generation with "bash generate.sh" succeeds.

Go build is now failing because we need to add contexts to our K8s API
calls, which is expected after the library version update.

Generate code with correct imports